src/kotatsu-news/pyproject.toml
[project]
name = "kotatsu-news"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "Amane Suzuki", email = "amane.suzu@gmail.com" }]
dependencies = [
    "slack-sdk>=3.31.0",
    "blockkit>=1.9.1",
    "python-dateutil>=2.9.0.post0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.5.7",
    "freezegun>=1.5.1",
    "pytest>=8.3.2",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/kotatsu_news"]
